+++ diagramming-tool/.gitignore
@@ -4,8 +4,10 @@
 /node_modules
 /.pnp
 .pnp.js
-
 package-lock.json
+
+# settings and dependencies
+settings.json
 package.json
 
 # testing
