+++ diagramming-tool/src/Component/Dashboard/Canvas.css
@@ -1,4 +1,4 @@
-.container {
+.dashboard-container {
   display: flex;
 }
 
@@ -20,7 +20,7 @@ h2 {
   font-size: 1.5rem;
 }
 
-button {
+.dashboard-container button {
   margin-right: 10px;
   padding: 10px;
   cursor: pointer;
@@ -35,7 +35,7 @@ button.selected {
 }
 
 button:hover {
-  background-color: #eb8080;
+  background-color: #f3e9e9;
 }
 
 .tooltip {
